requirements.txt


langchain
python-dotenv
langchain-openai
ipykernel
beautifulsoup4
faiss-cpu
langchain_groq
langchain_core
uvicorn
langserve
fastapi
sse_starlette
streamlit
pypdf
langchain_huggingface
langchain_core